+++ pkg/local_object_storage/metabase/control.go
@@ -6,13 +6,20 @@ import (
 	"path/filepath"
 	"slices"
 
+	"github.com/mr-tron/base58"
 	"github.com/nspcc-dev/bbolt"
 	bolterrors "github.com/nspcc-dev/bbolt/errors"
+	"github.com/nspcc-dev/neofs-node/pkg/local_object_storage/blobstor/common"
 	"github.com/nspcc-dev/neofs-node/pkg/local_object_storage/shard/mode"
 	"github.com/nspcc-dev/neofs-node/pkg/local_object_storage/util/logicerr"
 	"github.com/nspcc-dev/neofs-node/pkg/util"
+	"github.com/nspcc-dev/neofs-sdk-go/object"
+	oid "github.com/nspcc-dev/neofs-sdk-go/object/id"
 	"go.uber.org/zap"
 )
+
+// resyncBatchSize is the number of objects to be put to the metabase in a single batch during resync.
+const resyncBatchSize = 1000
 
 // ErrDegradedMode is returned when metabase is in a degraded mode.
 var ErrDegradedMode = logicerr.New("metabase is in a degraded mode")
@@ -210,4 +217,89 @@ func (db *DB) Reload(opts ...Option) (bool, error) {
 	}
 
 	return false, nil
+}
+
+// ResyncFromBlobstor resets the metabase and repopulates it by iterating
+// over all objects in the blobstor. shardID is written to the metabase after
+// the reset. onIterationError is called for every object that fails to be read
+// or unmarshalled; returning a non-nil error aborts the resync;
+// if onIterationError is nil, iteration errors are ignored.
+func (db *DB) ResyncFromBlobstor(bs common.Storage, shardID string, onIterationError func(oid.Address, error) error) error {
+	if onIterationError == nil {
+		onIterationError = func(oid.Address, error) error { return nil }
+	}
+
+	err := db.Reset()
+	if err != nil {
+		return fmt.Errorf("could not reset metabase: %w", err)
+	}
+
+	if shardID != "" {
+		id, err := base58.Decode(shardID)
+		if err != nil {
+			return fmt.Errorf("invalid shard ID: %w", err)
+		}
+		if err := db.WriteShardID(id); err != nil {
+			return fmt.Errorf("could not write shard ID: %w", err)
+		}
+	}
+
+	rh := &resyncHandler{
+		onError: onIterationError,
+		db:      db,
+		batch:   make([]*object.Object, 0, resyncBatchSize),
+	}
+	err = bs.Iterate(rh.handle, onIterationError)
+	if err != nil {
+		return fmt.Errorf("could not put objects to the meta from blobstor: %w", err)
+	}
+
+	// Flush any remaining objects in the batch
+	if err := rh.flush(); err != nil {
+		return fmt.Errorf("could not flush remaining objects to metabase: %w", err)
+	}
+
+	err = db.SyncCounters()
+	if err != nil {
+		return fmt.Errorf("could not sync object counters: %w", err)
+	}
+
+	return nil
+}
+
+// resyncHandler accumulates objects and flushes them in batches.
+type resyncHandler struct {
+	onError func(oid.Address, error) error
+	db      *DB
+	batch   []*object.Object
+}
+
+func (rh *resyncHandler) handle(addr oid.Address, data []byte) error {
+	obj := new(object.Object)
+
+	if err := obj.Unmarshal(data); err != nil {
+		return rh.onError(addr, err)
+	}
+
+	rh.batch = append(rh.batch, obj)
+
+	if len(rh.batch) == cap(rh.batch) {
+		return rh.flush()
+	}
+
+	return nil
+}
+
+func (rh *resyncHandler) flush() error {
+	if len(rh.batch) == 0 {
+		return nil
+	}
+
+	err := rh.db.PutBatch(rh.batch)
+	if err != nil {
+		return err
+	}
+
+	rh.batch = rh.batch[:0]
+	return nil
 }